+++ mycodo/mycodo_flask/routes_general.py
@@ -843,8 +843,13 @@ def computer_command(action):
             elif action == 'frontend_reload':
                 subprocess.Popen('docker restart mycodo_flask 2>&1', shell=True)
         else:
-            cmd = f'{INSTALL_DIRECTORY}/mycodo/scripts/mycodo_wrapper {action} 2>&1'
-            subprocess.Popen(cmd, shell=True)
+            if action == 'frontend_reload':
+                logger.info("Reloading frontend in 10 seconds")
+                cmd = f"sleep 10 && {INSTALL_DIRECTORY}/mycodo/scripts/mycodo_wrapper frontend_reload 2>&1"
+                subprocess.Popen(cmd, shell=True)
+            else:
+                cmd = f'{INSTALL_DIRECTORY}/mycodo/scripts/mycodo_wrapper {action} 2>&1'
+                subprocess.Popen(cmd, shell=True)
 
         if action == 'restart':
             flash(gettext("System rebooting in 10 seconds"), "success")
@@ -853,7 +858,7 @@ def computer_command(action):
         elif action == 'daemon_restart':
             flash(gettext("Command to restart the daemon sent"), "success")
         elif action == 'frontend_reload':
-            flash(gettext("Command to reload the frontend sent"), "success")
+            flash(gettext("Frontend reloading in 10 seconds"), "success")
 
         return redirect('/settings')