+++ mycodo/utils/tools.py
@@ -12,7 +12,7 @@ from collections import OrderedDict
 from dateutil import relativedelta
 
 from mycodo.config import (INSTALL_DIRECTORY, PATH_ACTIONS_CUSTOM,
-                           PATH_FUNCTIONS_CUSTOM, PATH_HTML_USER,
+                           PATH_FUNCTIONS_CUSTOM, PATH_TEMPLATE_USER,
                            PATH_INPUTS_CUSTOM, PATH_OUTPUTS_CUSTOM,
                            PATH_PYTHON_CODE_USER, PATH_USER_SCRIPTS,
                            PATH_WIDGETS_CUSTOM, SQL_DATABASE_MYCODO,
@@ -95,7 +95,7 @@ def create_settings_export(save_path=None):
                 (PATH_OUTPUTS_CUSTOM, "custom_outputs"),
                 (PATH_WIDGETS_CUSTOM, "custom_widgets"),
                 (PATH_USER_SCRIPTS, "user_scripts"),
-                (PATH_HTML_USER, "user_html"),
+                (PATH_TEMPLATE_USER, "user_html"),
                 (PATH_PYTHON_CODE_USER, "user_python_code")
             ]
             for each_backup in export_directories: